Create status_update.py
This is a simple Python Function that updates Product Status (active, archive or draft) based on Shopify Product ID. It is a useful function to import it in a bigger project with mass Status Changes (e.g. you could have a spreadsheet with thousands of products that you would like to change their status).